Added: Burning stack tracking
Properly tracks the number of burning stacks you have, when
dehyrdrated by a magi.

Also corrects for the GMCP changes where certain afflictions, like
burning, will show up as "burning (#)". For now, it just ignores
the (#) part, since it doesn't seem to give new gmcp events when
the # of stacks changes. Only when it is initially added or removed.