Fix double queuing tree tattoo
When using tree tattoo to cure out of a lock, we could potentially
send "curing queue add touch tree" twice in a row. This would result
in touching tree immediately, and then queueing it to be touched
again automatically on tree balance, when it is likely no longer
relevant to curing out of the lock.