Minor updates to some prompt tag colors
Updates the yellow health color to be less bright (gold).
And updates the max kai/shin to be a brighter green.

Hopefully nobody was really set on the old values, but if you were
feel free to change them back or add your own prompt tag colors for
personal use.