Remove prompt sub from pause/unpause
Remove the automatic prompt sub from pause/unpause commands. These
were presumably there because otherwise you might not get a prompt
after. But since they typically will enable/disable curing, there
will be a prompt anyway. If used in triggers, the promptsub was
causing it to delete the line that trigger it, even if not prompt.