Disbale cured AI check to fix affs disappearing
There is AI logic such that if serverside cures something that we
think should be lower priority than other affs we have, we clear
all the affs above the one that was cured (taking into account the
type of cure).

This may be useful in theory, but in practice, people don't really
illusion affs anymore, we should be relying on GMCP anyway, and I think
this is causing us to clear affs that we still have, resulting in not
shifting priorities correctly. Issue #9