Update Version to 1.3-beta-1
Updates the version for a 1.3-beta-1 release.

The primary changes for this release:
- Switch to pure GMCP aff tracking, using WunderSys Foundation
- Added some doqueue changes I've had locally for a long time now.
  The immediate impact is that curing out of locks now uses the do
  queue with priority 5 (e.g. fitness, dragonheal, restore). If you
  use the doqueue for your attacks with a greater # priority than 5,
  you can keep queueing your attacks without interrupting lock
  curing.
  There are probably some other smaller improvements made to the
  doqueue that I use, but may not be completely clean yet. Explore and
  use at your own risk.

Install Notes:
WunderSys now depends on WunderSys Foundation to function properly.
You may either install the WunderSys+Foundation package to install
both together, or you may install each xml separately. If installing
separately, it is recommended to install the Foundation first and make
sure it is listed above WunderSys in the scripts GUI.